Сложные функции вынесены в модули
