+++ src/mock/generate-description.js
@@ -1,0 +1,15 @@
+import {generateRandomIntegerFromRange, getRandomArrayElements} from './mock-utils';
+
+const SENTENCES_MAX_COUNT = 5;
+const RAW_TEXT_SOURCE = `Lorem ipsum dolor sit amet, consectetur adipiscing elit. Cras aliquet varius magna, non porta ligula feugiat eget. Fusce tristique felis at fermentum pharetra. Aliquam id orci ut lectus varius viverra. Nullam nunc ex, convallis sed finibus eget, sollicitudin eget ante. Phasellus eros mauris, condimentum sed nibh vitae, sodales efficitur ipsum. Sed blandit, eros vel aliquam faucibus, purus ex euismod diam, eu luctus nunc ante ut dui. Sed sed nisi sed augue convallis suscipit in sed felis. Aliquam erat volutpat. Nunc fermentum tortor ac porta dapibus. In rutrum ac purus sit amet tempus.`;
+
+const sentences = RAW_TEXT_SOURCE
+  .split(`.`)
+  .map((sentence) => sentence.trim())
+  .filter((str) => str);
+
+export const generateDescription = () => {
+  const desiredSentencesCount = generateRandomIntegerFromRange(1, SENTENCES_MAX_COUNT);
+
+  return getRandomArrayElements(sentences, desiredSentencesCount).join(`. `) + `.`;
+};
