+++ src/mock/mock-utils.js
@@ -1,3 +1,10 @@
-export const generateRandomIntegerFromRange = (lower = 0, upper = 1) => {
-  return lower + Math.floor(Math.random() * (upper - lower + 1));
-};
+import {generateRandomIntegerFromRange} from './utils/random-integer';
+import {generateName} from './utils/generate-name';
+import {getRandomArrayElements} from './utils/random-array-elements';
+import ArrayCyclicIterator from './utils/array-cyclic-iterator';
+
+const getRandomArrayElement = (arr) => arr[generateRandomIntegerFromRange(0, arr.length - 1)];
+
+const generateNames = (number) => Array(number).fill().map(generateName);
+
+export {generateRandomIntegerFromRange, generateName, generateNames, getRandomArrayElement, getRandomArrayElements, ArrayCyclicIterator};
